Add combinedReducers; Start initial books inside index.js instead of reducer file; Moved containers to containers folder
