+++ src/reducers/books.js
@@ -1,28 +1,6 @@
 import { CREATE_BOOK, REMOVE_BOOK } from '../actions/index';
 
-const initialState = [
-  {
-    title: 'The power of habit',
-    category: 'self improvement',
-    id: Math.floor(Math.random() * 1000),
-  },
-  {
-    title: 'The prince',
-    category: 'Politics',
-    id: Math.floor(Math.random() * 1000),
-  },
-  {
-    title: 'The everything store',
-    category: 'Biography',
-    id: Math.floor(Math.random() * 1000),
-  },
-  {
-    title: "The Hitchhiker's Guide to the Galaxy",
-    category: 'Sci-Fi',
-    id: Math.floor(Math.random() * 1000),
-  },
-];
-const books = (state = initialState, action) => {
+const books = (state = [], action) => {
   switch (action.type) {
     case CREATE_BOOK:
       return [...state, action.book];
