+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux';
+import books from './books';
+
+const rootReducer = combineReducers({ books });
+
+export default rootReducer;
